+++ src/shared/presets/project-filter-configs.ts
@@ -10,10 +10,6 @@ export const SINGLE_SELECT_OPTIONS = {
     { label: "Has been hacked", value: true },
     { label: "Has not been hacked", value: false },
   ],
-  audits: [
-    { label: "Has been audited", value: true },
-    { label: "Has not been audited", value: false },
-  ],
   mainNet: [
     { label: "Deployed on Mainnet", value: true },
     { label: "Not Deployed on Mainnet", value: false },
@@ -111,11 +107,10 @@ export const FILTER_CONFIG_PRESETS = {
   audits: {
     position: 9,
     label: "Audits",
-    show: false,
+    show: true,
     googleAnalyticsEventName: null,
     googleAnalyticsEventId: null,
-    kind: FilterKind.RANGE,
-    options: SINGLE_SELECT_OPTIONS.audits,
+    kind: FilterKind.MULTI_SELECT_WITH_SEARCH,
   },
   hacks: {
     position: 10,